requirements.txt
flask
flask-cors
pycryptodome
beautifulsoup4
ddddocr
requests
aligo
websockets